#0 - abtraktní třída pro commandy
