+++ app/Check/Commands/PublishAliveChecksCommand.php
@@ -2,55 +2,16 @@
 
 namespace Pd\Monitoring\Check\Commands;
 
-class PublishAliveChecksCommand extends \Symfony\Component\Console\Command\Command
+use Pd\Monitoring\Commands\TNamedCommand;
+
+class PublishAliveChecksCommand extends PublishChecksCommand
 {
+	use TNamedCommand;
 
-	use \Pd\Monitoring\Commands\TNamedCommand;
-
-	/**
-	 * @var \Kdyby\RabbitMq\IProducer
-	 */
-	private $producer;
-
-	/**
-	 * @var \Pd\Monitoring\Check\ChecksRepository
-	 */
-	private $checksRepository;
-
-
-	public function __construct(
-		\Kdyby\RabbitMq\IProducer $producer,
-		\Pd\Monitoring\Check\ChecksRepository $checksRepository
-	) {
-		parent::__construct();
-
-		$this->producer = $producer;
-		$this->checksRepository = $checksRepository;
-	}
-
-
-	protected function configure()
+	protected function getConditions(): array
 	{
-		parent::configure();
-
-		$this->setName($this->generateName());
-	}
-
-
-	protected function execute(
-		\Symfony\Component\Console\Input\InputInterface $input,
-		\Symfony\Component\Console\Output\OutputInterface $output
-	) {
-		$conditions = [
+		return [
 			'type' => \Pd\Monitoring\Check\ICheck::TYPE_ALIVE,
 		];
-		$checks = $this->checksRepository->findAll($conditions);
-
-		foreach ($checks as $check) {
-			$this->producer->publish($check->id);
-		}
-
-		return 0;
 	}
-
 }
